+++ src/test/java/kitchenpos/fixture/MenuProductDomainFixture.java
@@ -1,10 +1,8 @@
-package kitchenpos.menu.domain.fixture;
+package kitchenpos.fixture;
 
 import kitchenpos.menu.domain.menu.Menu;
 import kitchenpos.menu.domain.menuproduct.MenuProduct;
 import kitchenpos.menu.domain.product.Product;
-
-import java.math.BigDecimal;
 
 public class MenuProductDomainFixture {
     public static MenuProduct menuProduct(Menu menu, Product product, int quantity) {
