+++ src/test/java/kitchenpos/menu/domain/menugroup/MenuGroupTest.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
-import static kitchenpos.menu.domain.fixture.MenuGroupDomainFixture.menuGroup;
+import static kitchenpos.fixture.MenuGroupDomainFixture.menuGroup;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
