+++ src/test/java/kitchenpos/menu/domain/product/ProductMockitoTest.java
@@ -10,7 +10,7 @@ import org.junit.jupiter.api.Test;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import static kitchenpos.menu.domain.fixture.ProductDomainFixture.*;
+import static kitchenpos.fixture.ProductDomainFixture.*;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.mockito.Mockito.*;
