+++ src/test/java/kitchenpos/menu/domain/product/ProductTest.java
@@ -7,7 +7,7 @@ import org.junit.jupiter.api.Test;
 
 import java.math.BigDecimal;
 
-import static kitchenpos.menu.domain.fixture.ProductDomainFixture.product;
+import static kitchenpos.fixture.ProductDomainFixture.product;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
