+++ src/test/java/kitchenpos/table/domain/table/TableMockitoTest.java
@@ -13,7 +13,7 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
 
-import static kitchenpos.menu.domain.fixture.OrderTableDomainFixture.*;
+import static kitchenpos.fixture.OrderTableDomainFixture.*;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.*;
 import static org.mockito.Mockito.mock;
